Make nested adj dicts lockable
